Update Valid BST.cpp

+++ Trees/BST/Valid BST.cpp
@@ -1,7 +1,8 @@
+
 class Solution {
 public:
     bool isValidBST(TreeNode* root) {
-        return isValidBST(root, INT_MIN, INT_MAX);
+        return isValidBST(root, LONG_MIN, LONG_MAX);
     }
     bool isValidBST(TreeNode* root, long minval , long maxval){
         if(root==nullptr) return true;
